Divid the ListItem module info submodules
